refactor (Lotto): 로또 사이즈 검증 로직 메소드로 분리
